--- Merging r22401 into '.': U    packages/fcl-db/tests/dbtestframework.pas U    packages/fcl-db/tests/dbtestframework_gui.lpi U    packages/fcl-db/tests/database.ini.txt U    packages/fcl-db/tests/dbtestframework_gui.lpr U    packages/fcl-db/tests/tcsdfdata.pp U    packages/fcl-db/tests/toolsunit.pas U    packages/fcl-db/tests/sdfdstoolsunit.pas U    packages/fcl-db/tests/README.txt --- Merging r22470 into '.': U    packages/fcl-db/src/sqldb/postgres/pqconnection.pp --- Merging r22559 into '.': C    packages/fcl-db/fpmake.pp A    packages/fcl-db/src/sqlite/sqlite3backup.pas --- Merging r22560 into '.': G    packages/fcl-db/tests/README.txt --- Merging r22583 into '.': G    packages/fcl-db/fpmake.pp D    packages/fcl-db/src/sqlite/sqlite3backup.pas A    packages/fcl-db/src/sqldb/sqlite/sqlite3backup.pas --- Merging r22609 into '.': U    packages/fcl-db/src/README.txt Summary of conflicts:   Text conflicts: 1
# revisions: 22401,22470,22559,22560,22583,22609
r22401 | michael | 2012-09-16 13:38:49 +0200 (Sun, 16 Sep 2012) | 1 line
Changed paths:
   M /trunk/packages/fcl-db/tests/README.txt
   M /trunk/packages/fcl-db/tests/database.ini.txt
   M /trunk/packages/fcl-db/tests/dbtestframework.pas
   M /trunk/packages/fcl-db/tests/dbtestframework_gui.lpi
   M /trunk/packages/fcl-db/tests/dbtestframework_gui.lpr
   M /trunk/packages/fcl-db/tests/sdfdstoolsunit.pas
   M /trunk/packages/fcl-db/tests/tcsdfdata.pp
   M /trunk/packages/fcl-db/tests/toolsunit.pas

* Applied patch from #22899
r22470 | michael | 2012-09-27 11:10:08 +0200 (Thu, 27 Sep 2012) | 1 line
Changed paths:
   M /trunk/packages/fcl-db/src/sqldb/postgres/pqconnection.pp

* Patch to add support for charset (bug 22985)
r22559 | marco | 2012-10-06 14:29:36 +0200 (Sat, 06 Oct 2012) | 3 lines
Changed paths:
   M /trunk/packages/fcl-db/fpmake.pp
   A /trunk/packages/fcl-db/src/sqlite/sqlite3backup.pas

 * Sqlite3backup, implements backups between 2 TSQLite3Connections, TSQLite3Connection backup to file and TSQLite3Connection restore from file. 
    by Ludo Brands, mantis #22394
r22560 | marco | 2012-10-06 14:31:15 +0200 (Sat, 06 Oct 2012) | 2 lines
Changed paths:
   M /trunk/packages/fcl-db/tests/README.txt

 * patch from Reinier: More details on setting up test database, ignoring tests for certain databases.... mantis #22970
r22583 | ludob | 2012-10-08 11:09:27 +0200 (Mon, 08 Oct 2012) | 1 line
Changed paths:
   M /trunk/packages/fcl-db/fpmake.pp
   A /trunk/packages/fcl-db/src/sqldb/sqlite/sqlite3backup.pas
   D /trunk/packages/fcl-db/src/sqlite/sqlite3backup.pas

Moved sqlite3backup.pas to src/sqldb/sqlite
r22609 | reiniero | 2012-10-10 09:45:56 +0200 (Wed, 10 Oct 2012) | 1 line
Changed paths:
   M /trunk/packages/fcl-db/src/README.txt

fcl-db readme: sdfdataset is meant to be RFC4180 CSV compliant; list of dbs updated; typos fixed.

git-svn-id: branches/fixes_2_6@22702 -

+++ packages/fcl-db/tests/sdfdstoolsunit.pas
@@ -46,11 +46,14 @@ end;
 procedure TSdfDSDBConnector.CreateNDatasets;
 var countID,n : integer;
 begin
+  if dbname='' then raise Exception.Create('dbname variable not specified. You must specify name= in database.ini');
   for n := 0 to MaxDataSet do
     begin
     with TSdfDataSet.Create(nil) do
       begin
       FileName := dbname+PathDelim+'fpdev_'+inttostr(n)+'.dat';
+      // Make sure the directory exists so we can write
+      ForceDirectories(dbname);
       DeleteFile(FileName);
       FileMustExist:=False;
       
@@ -78,10 +81,13 @@ end;
 procedure TSdfDSDBConnector.CreateFieldDataset;
 var i : integer;
 begin
+  if dbname='' then raise Exception.Create('dbname variable not specified. You must specify name= in database.ini');
   with TSdfDataSet.Create(nil) do
     begin
     FileName := dbname+PathDelim+'fpdev_field.dat';
       DeleteFile(FileName);
+    // Make sure the directory exists so we can write
+    ForceDirectories(dbname);
     FileMustExist:=False;
     
     SetFieldDatasetSchema(Schema);
